examples/python/requirements.txt
pandas
pyarrow
polars
qdrant-client
openai
anthropic
matplotlib
plotly
openpyxl
python-dotenv
langchain
langchain-openai
